Store menu items in memory to optimize the popup time
Previously, all menu items were built from the history when the
menu was opened. This was an easy solution because we didn't need to
maintain the state and order of menu items. However, this is getting
slow if we have a long history.

The commit changes the architecture to build all the menu items once,
keep them in memory and synchronize with the history. It's more
cumbersome and it's definitely far from the code I want myself to write,
but it does the job with some notable underlying changes:

1. Sorter class was reworked to return in descending order.
2. Search class was reworked to operate on array of strings.
3. Additional observers are required for configuration which changes an
   order or size of kept memory. For example, when sortBy is changed,
   the whole list of menu items needs to be re-sorted.

Related to #113.