Allow to paste without formatting
The new menu item which copies, clears formatting and pastes item is
added. It's general key modifier is SHIFT and is paired with pasting
key modifier. This means that by default, one needs to use OPTION+SHIFT
to paste without formatting. If pasting is enabled by default, the
keybinding changes to COMMAND+SHIFT.

There is also a new setting to remove formatting by default when
pasting. The preferences window also includes a comprehensive
explanation of the possible keybinding combinations which depend on the
pasteByDefault and removeFormattingByDefault settings.

Removal of formatting is implemented by copying data from
NSPasteboard.PasteboardType.string of the item.

Closes #109.