Enable application sandbox
This enables a sandbox which is more secure and is required for App
Store submission. The application doesn't use any restricted features so
it's only required to properly migrate preferences (this is done
automatically) and storage (this is done by moving Application Support
directory using special container-migration manifest). See
https://developer.apple.com/library/archive/documentation/Security/Conceptual/AppSandboxDesignGuide/MigratingALegacyApp/MigratingAnAppToASandbox.html
for more details.

In order to keep auto-update work, Sparkle needed to be updated to 2.x
as 1.x doesn't work in sandboxed environment. It's not possible to
install 2.x using CocoaPods so it's now embedded as a framework. It has
been build in 2.x branch at 438f4a21. Downloading and installing updates
is implemented through XPC services to avoid adding unnecessary network
entitlements to the application sandbox. See
https://github.com/sparkle-project/Sparkle/blob/2.x/INSTALL.markdown for
more details.