Avoid highlighting search field when there are no search results
This will keep the focus on the search field when there are no search
results allowing the user to keep typing or clear the field by pressing
backspace. It also prevents tiny annoyance that when there are no search
results, the search field is selected with all its content highlighted
like if someone would press Cmd+A on it.