Add basic support for shortcuts
This commits adds two shortcut actions that can be used to automate Maccy.

The first action selects Nth history item in the clipboard history
similary to user clicking it manually. This allows to for example assign
a global hotkey to automatically select the previously copied item and
paste it. Likewise, it can be used to assign hotkeys for selecting any
history item knowing its positiont number. This implements a solution
for advanced usage needed in #315. The shortcut returns the selected
item title as an output which allows sending it to other actions (e.g.
show notification with the selected item title).

The second action simply clears the history and allows to automate #340.