Update all packages to recent releases
Most of the updates are seamless, though switching to modern
LaunchAtLogin forces to update minimum macOS version to 13. That's fine
as long as we're still going to use SwiftData (requires macOS 14).

KeyboardShortcuts no longer work in settings UI, but it's okay as long
as the next commits will re-write settings to SwiftUI and fix this
issue.

fuse-swift was not updated and it's actually archived, so a suitable
replacement needs to be found eventually.

The commit also clears all migrations for UserDefaults that accumulated
over 7 years. It was necessary to avoid update migrations for changed
APIs in packages.