Fix pasting from panel opened via menu icon
There is a huge problem when using MenuBarExtra with .window style. It creates an extra window that cannot be reliably hidden or taken out of the responder chain. So when the pasting is attempted from the main window, the paste key actually goes to this extra window.

Initial attempt to hide this window using MenuBarExtraAccess library didn't help. Until Apple provides a way to assign custom windows or scenes to menu bar extra apps, we have to resort to using AppDelegate and AppKit for this purpose. It's much more reliable and predictable.

This commit also fixes a bug when selecting an item with Option pressed within a popup opened via menu icon would turn off Maccy. This is due to the fact there is no way to limit the click to the bounds of the menu icon button.